src/requirements.txt
paho-mqtt==1.6.1
influxdb-client
torch
scikit-learn
numpy
simple-pid==2.0.0
